+++ admin/metabox/ajax/sanitizers/subforms/class-sanitize-links.php
@@ -3,7 +3,7 @@
  * Registers the Sanitize_Links class.
  *
  * @package Style_Blocks\Sanitizer
- * @since 2.1.0
+ * @since 3.0.0
  */
 
 namespace Style_Blocks;
@@ -14,7 +14,7 @@ namespace Style_Blocks;
  * Checks for and sanitizes the fields expected by the link.
  *
  * @package Style_Blocks\Sanitizer
- * @since 2.1.0
+ * @since 3.0.0
  */
 class Sanitize_Links {