+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _This sections lists changes committed since most recent release_
 ### Added:
 
 - Migration tool to detect and convert/delete blocks stored under the legacy data architecture
+- Link List block type to display a list of user specified links (currently disabled)
 
 ### Changed: